refactor: extract Deploy, CI/CD, K8s, Plugin, Registry, System services (Phase 4)
Agent-Task: bridge-refactor-phase4
Agent-Model: SOLO
Agent-Decision: Extract 6 domain service files to reduce bridge.go from 1734 to 448 lines

Bridge God Object decomposition COMPLETE:
- bridge.go: 2986 → 448 lines (-85%)
- 87 methods distributed across 18 files

Phase 4 extracted services:
- deploy_service.go (601 lines): Deploy, DeployAsync, BuildAndDeploy, BatchDeploy,
  BatchDeployWithConfig, Rollback, RollbackWithType, GetContainerStatus, Stop,
  Remove, GetContainerLogs, GetDeploymentHistory, saveDeploymentRecord,
  GetLatestDeploymentRecord, findPreviousSuccessfulImage, buildRollbackConfig,
  saveRollbackRecord
- cicd_service.go (130 lines): TriggerCIBuild, GetCIBuildStatus
- k8s_service.go (137 lines): K8sDeploy, K8sListDeployments, K8sGetPods
- plugin_service.go (146 lines): PluginOps, ListPlugins, GetPluginInfo
- registry_service.go (128 lines): RegistryOps
- system_service.go (182 lines): DetectEnv, HealthCheck, CheckDeployReadiness,
  PerformSystemUpdate, HealContainer

All methods remain on Bridge struct receiver to satisfy mcp.Deployer interface.
Zero behavioral changes - pure code relocation.
