+++ internal/service/bridge.go
@@ -370,781 +370,18 @@ func (is *interactiveSession) Close() error {
 	return is.session.Close()
 }
 
-// ---------- 1. Deploy ----------
-
-// DeployAsync starts a deploy in a goroutine and returns a task ID immediately.
-func (b *Bridge) DeployAsync(ctx context.Context, cfg mcp.DeployConfig, appID string) (taskID string, err error) {
-	taskID = createTask("deploy")
-	traceID := tracing.TraceIDFromContext(ctx)
-	updateTask(taskID, "running", 0, "deploy started")
-	go func() {
-		cs, deployErr := b.Deploy(ctx, cfg)
-		if deployErr != nil {
-			updateTask(taskID, "failed", 100, deployErr.Error())
-			b.EventBus.Publish(DeployEvent{
-				TaskID:    taskID,
-				AppID:     appID,
-				Step:      "done",
-				Status:    "failed",
-				Progress:  100,
-				Message:   deployErr.Error(),
-				Timestamp: time.Now().Format(time.RFC3339),
-				TraceID:   traceID,
-			})
-		} else {
-			updateTask(taskID, "success", 100, "deploy completed")
-			b.EventBus.Publish(DeployEvent{
-				TaskID:    taskID,
-				AppID:     appID,
-				Step:      "done",
-				Status:    "success",
-				Progress:  100,
-				Message:   "deploy completed",
-				Timestamp: time.Now().Format(time.RFC3339),
-				TraceID:   traceID,
-			})
-			taskMu.Lock()
-			if t, ok := tasks[taskID]; ok {
-				t.Result = cs
-			}
-			taskMu.Unlock()
-		}
-	}()
-	return taskID, nil
-}
-
-func (b *Bridge) Deploy(ctx context.Context, cfg mcp.DeployConfig) (*mcp.ContainerStatus, error) {
-	// Record deploy start time for metrics
-	deployStart := time.Now()
-
-	// Determine executor: remote SSH if server_id provided, otherwise local
-	executor := b.Executor
-	var host string
-	var port int
-
-	if cfg.ServerID != "" {
-		// Look up server info for preflight
-		row := make(map[string]interface{})
-		if err := b.DB.Table("servers").Where("id = ?", cfg.ServerID).Take(&row).Error; err != nil {
-			return nil, fmt.Errorf("server not found: %w", err)
-		}
-		host = toString(row["host"])
-		port = toInt(row["port"])
-
-		remoteExec, err := b.getRemoteExecutor(ctx, cfg.ServerID)
-		if err != nil {
-			return nil, fmt.Errorf("failed to get remote executor for server %s: %w", cfg.ServerID, err)
-		}
-		defer func() {
-			if cerr := remoteExec.Close(); cerr != nil {
-				slog.Warn("failed to close remote executor", "error", cerr)
-			}
-		}()
-		executor = remoteExec
-	}
-
-	// Run preflight checks
-	pfCfg := PreflightConfig{
-		Host:         host,
-		Port:         port,
-		Executor:     executor,
-		PortMappings: cfg.Ports,
-	}
-	pfResult := RunPreflight(ctx, pfCfg)
-	if !pfResult.Passed {
-		// Save preflight failure to deployment records
-		b.saveDeploymentRecord(ctx, mcp.DeployConfig{
-			Image:         cfg.Image,
-			ContainerName: cfg.ContainerName,
-			ServerID:      cfg.ServerID,
-		}, "preflight_failed", pfResult)
-
-		// Log structured preflight failure
-		logPreflightResult(cfg.ContainerName, pfResult)
-
-		b.EventBus.Publish(DeployEvent{
-			TaskID:    "", AppID: cfg.ContainerName,
-			Step: "preflight", Status: "failed", Progress: 20,
-			Message:   pfResult.Message,
-			Timestamp: time.Now().Format(time.RFC3339),
-		})
-
-		// Record preflight failure metric
-		metrics.DeployTotal.WithLabelValues(cfg.ContainerName, cfg.ServerID, "failed").Inc()
-		metrics.DeployDuration.Observe(time.Since(deployStart).Seconds())
-
-		return nil, &PreflightError{
-			Code:    pfResult.Code,
-			Message: pfResult.Message,
-			Checks:  pfResult.Checks,
-		}
-	}
-
-	b.EventBus.Publish(DeployEvent{
-		TaskID:    "", AppID: cfg.ContainerName,
-		Step: "preflight", Status: "success", Progress: 20,
-		Message:   "preflight checks passed",
-		Timestamp: time.Now().Format(time.RFC3339),
-	})
-
-	dCfg := deployer.DeployConfig{
-		Image:         cfg.Image,
-		ContainerName: cfg.ContainerName,
-		Ports:         cfg.Ports,
-		EnvVars:       cfg.EnvVars,
-		RestartPolicy: cfg.RestartPolicy,
-		Network:       cfg.Network,
-		Volumes:       cfg.Volumes,
-		Labels:        cfg.Labels,
-		ResourceLimits: deployer.ResourceLimits{
-			CPU:    cfg.CPU,
-			Memory: cfg.Memory,
-		},
-	}
-
-	d := deployer.New(executor)
-
-	b.EventBus.Publish(DeployEvent{
-		TaskID:    "", AppID: cfg.ContainerName,
-		Step: "pull", Status: "running", Progress: 30,
-		Message:   "pulling image: " + cfg.Image,
-		Timestamp: time.Now().Format(time.RFC3339),
-	})
-
-	cs, err := d.Deploy(ctx, dCfg)
-	if err != nil {
-		// Save deployment failure
-		b.saveDeploymentRecord(ctx, cfg, "failed", nil)
-		slog.Error("container deployment failed", "container", cfg.ContainerName, "error", err)
-
-		// Record deploy failure metric
-		metrics.DeployTotal.WithLabelValues(cfg.ContainerName, cfg.ServerID, "failed").Inc()
-		metrics.DeployDuration.Observe(time.Since(deployStart).Seconds())
-
-		b.EventBus.Publish(DeployEvent{
-			TaskID:    "", AppID: cfg.ContainerName,
-			Step: "run", Status: "failed", Progress: 60,
-			Message:   "deploy failed: " + err.Error(),
-			Timestamp: time.Now().Format(time.RFC3339),
-		})
-
-		return nil, err
-	}
-	// Save deployment success
-	b.saveDeploymentRecord(ctx, cfg, "success", nil)
-	slog.Info("container deployed successfully", "container", cfg.ContainerName, "container_id", cs.ID)
-
-	// Record deploy success metric
-	metrics.DeployTotal.WithLabelValues(cfg.ContainerName, cfg.ServerID, "success").Inc()
-	metrics.DeployDuration.Observe(time.Since(deployStart).Seconds())
-
-	b.EventBus.Publish(DeployEvent{
-		TaskID:    "", AppID: cfg.ContainerName,
-		Step: "run", Status: "success", Progress: 90,
-		Message:   "container deployed successfully",
-		Timestamp: time.Now().Format(time.RFC3339),
-	})
-
-	return &mcp.ContainerStatus{
-		ID:        cs.ID,
-		Name:      cs.Name,
-		Image:     cs.Image,
-		Status:    cs.Status,
-		Ports:     cs.Ports,
-		CreatedAt: cs.CreatedAt.Format(time.RFC3339),
-		Labels:    cs.Labels,
-	}, nil
-}
-
-// ---------- 1b. BuildAndDeploy ----------
-
-// BuildAndDeploy orchestrates the full build-and-deploy pipeline:
-// git clone -> detect tech stack -> generate Dockerfile -> docker build -> deploy.
-func (b *Bridge) BuildAndDeploy(ctx context.Context, cfg mcp.BuildAndDeployConfig) (*mcp.BuildAndDeployResult, error) {
-	exec := b.Executor
-	if exec == nil {
-		return nil, fmt.Errorf("no executor available")
-	}
-
-	// Convert MCP config to builder config
-	bldCfg := builder.BuildConfig{
-		RepoURL:             cfg.RepoURL,
-		Branch:              cfg.Branch,
-		TechStack:           cfg.TechStack,
-		AppName:             cfg.AppName,
-		ProjectDir:          cfg.ProjectDir,
-		BuildArgs:           cfg.BuildArgs,
-		EnvVars:             cfg.EnvVars,
-		Ports:               cfg.Ports,
-		ServerID:            cfg.ServerID,
-		DockerfileOverrides: cfg.DockerfileOverrides,
-		RegistryID:          cfg.RegistryID,
-		PushImage:           cfg.PushImage,
-		ImageTag:            cfg.ImageTag,
-	}
-
-	bld := builder.NewBuilder(exec)
-	result, err := bld.BuildAndDeploy(ctx, bldCfg)
-	if err != nil {
-		return nil, err
-	}
-
-	// After successful build, deploy the built image
-	deployCfg := mcp.DeployConfig{
-		Image:         result.Image,
-		ContainerName: cfg.AppName,
-		Ports:         cfg.Ports,
-		EnvVars:       cfg.EnvVars,
-	}
-	if cfg.ServerID != "" {
-		deployCfg.ServerID = cfg.ServerID
-	}
-
-	_, err = b.Deploy(ctx, deployCfg)
-	if err != nil {
-		return nil, fmt.Errorf("build succeeded but deploy failed: %w", err)
-	}
-
-	// Convert builder result to MCP result
-	return &mcp.BuildAndDeployResult{
-		Image:      result.Image,
-		Digest:     result.Digest,
-		Size:       result.Size,
-		BuildLog:   result.BuildLog,
-		Duration:   result.Duration,
-		TechStack:  result.TechStack,
-		CommitHash: result.CommitHash,
-	}, nil
-}
-
-// ---------- 2. GetContainerStatus ----------
-
-func (b *Bridge) GetContainerStatus(ctx context.Context, name string) (*mcp.ContainerStatus, error) {
-	cs, err := b.d().GetContainerStatus(ctx, name)
-	if err != nil {
-		return nil, err
-	}
-
-	return &mcp.ContainerStatus{
-		ID:        cs.ID,
-		Name:      cs.Name,
-		Image:     cs.Image,
-		Status:    cs.Status,
-		Ports:     cs.Ports,
-		CreatedAt: cs.CreatedAt.Format(time.RFC3339),
-		Labels:    cs.Labels,
-	}, nil
-}
 
 // ---------- 3. ListApps ----------
 
 
-// ---------- 7. Stop ----------
-
-func (b *Bridge) Stop(ctx context.Context, name string) error {
-	return b.d().Stop(ctx, name)
-}
-
-// ---------- 8. Remove ----------
-
-func (b *Bridge) Remove(ctx context.Context, name string) error {
-	return b.d().Remove(ctx, name)
-}
-
-// ---------- 9. GetContainerLogs ----------
-
-func (b *Bridge) GetContainerLogs(ctx context.Context, name string, tail int) (string, error) {
-	return b.d().GetContainerLogs(ctx, name, tail)
-}
-
-// ---------- 10. Rollback ----------
-
-// RollbackResult contains detailed information about a rollback operation.
-type RollbackResult struct {
-	ContainerStatus *mcp.ContainerStatus `json:"container_status"`
-	PreviousImage   string               `json:"previous_image"`
-	TargetImage     string               `json:"target_image"`
-	DeployType      string               `json:"deploy_type"` // rollback or auto_rollback
-}
-
-// Rollback reverts a container to a previous image version with full config preservation.
-// If previousImage is empty, it automatically finds the last successful deployment image.
-// The rollback preserves ports, env vars, volumes, network, labels, and resource limits.
-func (b *Bridge) Rollback(ctx context.Context, containerName, previousImage string) (*mcp.ContainerStatus, error) {
-	return b.RollbackWithType(ctx, containerName, previousImage, "rollback")
-}
-
-// RollbackWithType performs a rollback with a specified deploy type (rollback or auto_rollback).
-func (b *Bridge) RollbackWithType(ctx context.Context, containerName, previousImage, deployType string) (*mcp.ContainerStatus, error) {
-	if deployType == "" {
-		deployType = "rollback"
-	}
-
-	// Step 1: Resolve the target image — auto-find if not specified
-	if previousImage == "" {
-		found, err := b.findPreviousSuccessfulImage(ctx, containerName)
-		if err != nil {
-			return nil, fmt.Errorf("no previous image found for rollback: %w", err)
-		}
-		previousImage = found
-		slog.Info("rollback: auto-resolved previous image", "container", containerName, "image", previousImage)
-	}
-
-	// Step 2: Get current running image for recording
-	currentImage := ""
-	if cs, err := b.d().GetContainerStatus(ctx, containerName); err == nil {
-		currentImage = cs.Image
-	}
-
-	// Step 3: Build full deploy config from app record (preserves all settings)
-	cfg, err := b.buildRollbackConfig(ctx, containerName, previousImage)
-	if err != nil {
-		return nil, fmt.Errorf("failed to build rollback config: %w", err)
-	}
-
-	// Step 4: Atomic rollback — pull new image FIRST, then stop/remove old container
-	// This ensures we don't destroy the running container if the image pull fails
-	slog.Info("rollback: pulling target image", "container", containerName, "image", previousImage)
-	pullCmd := fmt.Sprintf("docker pull %s", previousImage)
-	if _, err := b.Executor.RunCommand(ctx, pullCmd); err != nil {
-		return nil, fmt.Errorf("rollback failed: could not pull image %s: %w", previousImage, err)
-	}
-
-	// Step 5: Stop and remove the current container
-	if err := b.d().Stop(ctx, containerName); err != nil {
-		slog.Warn("rollback: stop warning", "error", err)
-	}
-	if err := b.d().Remove(ctx, containerName); err != nil {
-		slog.Warn("rollback: remove warning", "error", err)
-	}
-
-	// Step 6: Deploy with full config
-	cs, err := b.Deploy(ctx, cfg)
-	if err != nil {
-		// Record failed rollback
-		b.saveRollbackRecord(ctx, cfg, currentImage, deployType, "failed", err.Error())
-		return nil, fmt.Errorf("rollback deploy failed: %w", err)
-	}
-
-	// Step 7: Record successful rollback
-	b.saveRollbackRecord(ctx, cfg, currentImage, deployType, "success", "")
-
-	// Step 8: Update app's current version
-	if cfg.AppName != "" {
-		b.DB.Table("apps").Where("id = ? OR name = ?", cfg.AppName, cfg.AppName).
-			Update("current_version", previousImage)
-	}
-
-	slog.Info("rollback completed", "container", containerName,
-		"from", currentImage, "to", previousImage, "type", deployType)
-
-	return cs, nil
-}
-
-// findPreviousSuccessfulImage looks up the most recent successful deployment record
-// for a container and returns its image.
-func (b *Bridge) findPreviousSuccessfulImage(ctx context.Context, containerName string) (string, error) {
-	if b.DB == nil {
-		return "", fmt.Errorf("database not available")
-	}
-
-	var record model.DeploymentRecord
-	err := b.DB.Where("container_name = ? AND status = ?", containerName, "success").
-		Order("created_at DESC").
-		Limit(1).
-		First(&record).Error
-	if err != nil {
-		return "", fmt.Errorf("no successful deployment found for container %s", containerName)
-	}
-
-	return record.Image, nil
-}
-
-// buildRollbackConfig constructs a full DeployConfig for rollback by reading the app record.
-// This ensures ports, env vars, resource limits, etc. are preserved during rollback.
-func (b *Bridge) buildRollbackConfig(ctx context.Context, containerName, targetImage string) (mcp.DeployConfig, error) {
-	cfg := mcp.DeployConfig{
-		Image:         targetImage,
-		ContainerName: containerName,
-		RestartPolicy: "unless-stopped",
-	}
-
-	if b.DB == nil {
-		return cfg, nil
-	}
-
-	// Look up the app by container_name or name
-	var app model.App
-	err := b.DB.Where("container_name = ? OR name = ?", containerName, containerName).First(&app).Error
-	if err != nil {
-		slog.Debug("rollback: no app record found, using minimal config", "container", containerName)
-		return cfg, nil
-	}
-
-	cfg.AppName = app.Name
-
-	// Restore ports from app record
-	if app.Domain != "" {
-		cfg.Ports = "80:80, 443:443"
-	}
-
-	// Restore env vars
-	if app.EnvVars != "" {
-		var envMap map[string]string
-		if json.Unmarshal([]byte(app.EnvVars), &envMap) == nil {
-			cfg.EnvVars = envMap
-		}
-	}
-
-	// Restore resource limits
-	if app.ResourceLimits != "" {
-		var limits map[string]string
-		if json.Unmarshal([]byte(app.ResourceLimits), &limits) == nil {
-			cfg.CPU = limits["cpu"]
-			cfg.Memory = limits["memory"]
-		}
-	}
-
-	// Try to get the config snapshot from the last successful deployment
-	var lastRecord model.DeploymentRecord
-	if err := b.DB.Where("container_name = ? AND status = ?", containerName, "success").
-		Order("created_at DESC").Limit(1).First(&lastRecord).Error; err == nil {
-		if lastRecord.ConfigSnapshot != "" {
-			var snapCfg mcp.DeployConfig
-			if json.Unmarshal([]byte(lastRecord.ConfigSnapshot), &snapCfg) == nil {
-				// Merge snapshot config (takes precedence over app record)
-				if snapCfg.Ports != "" {
-					cfg.Ports = snapCfg.Ports
-				}
-				if len(snapCfg.EnvVars) > 0 {
-					cfg.EnvVars = snapCfg.EnvVars
-				}
-				if snapCfg.Network != "" {
-					cfg.Network = snapCfg.Network
-				}
-				if snapCfg.Volumes != "" {
-					cfg.Volumes = snapCfg.Volumes
-				}
-				if len(snapCfg.Labels) > 0 {
-					cfg.Labels = snapCfg.Labels
-				}
-				if snapCfg.CPU != "" {
-					cfg.CPU = snapCfg.CPU
-				}
-				if snapCfg.Memory != "" {
-					cfg.Memory = snapCfg.Memory
-				}
-				if snapCfg.RestartPolicy != "" {
-					cfg.RestartPolicy = snapCfg.RestartPolicy
-				}
-			}
-		}
-	}
-
-	return cfg, nil
-}
-
-// saveRollbackRecord persists a rollback operation to the deployment records.
-func (b *Bridge) saveRollbackRecord(ctx context.Context, cfg mcp.DeployConfig, currentImage, deployType, status, errMsg string) {
-	if b.DB == nil {
-		return
-	}
-
-	// Serialize config snapshot
-	snapshotJSON, _ := json.Marshal(cfg)
-
-	record := &model.DeploymentRecord{
-		ID:             generateID(),
-		TenantID:       "tenant-default",
-		ServerID:       cfg.ServerID,
-		AppName:        cfg.AppName,
-		ContainerName:  cfg.ContainerName,
-		Image:          cfg.Image,
-		PreviousImage:  currentImage,
-		DeployType:     deployType,
-		ConfigSnapshot: string(snapshotJSON),
-		Status:         status,
-		ErrorMessage:   errMsg,
-		CreatedAt:      time.Now(),
-		UpdatedAt:      time.Now(),
-	}
-	if err := b.DB.Create(record).Error; err != nil {
-		slog.Error("failed to save rollback record", "error", err)
-	}
-}
-
-// GetDeploymentHistory returns the deployment history for a container, ordered by most recent first.
-func (b *Bridge) GetDeploymentHistory(ctx context.Context, containerName string, limit int) ([]model.DeploymentRecord, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-	if limit <= 0 || limit > 100 {
-		limit = 20
-	}
-
-	var records []model.DeploymentRecord
-	err := b.DB.Where("container_name = ?", containerName).
-		Order("created_at DESC").
-		Limit(limit).
-		Find(&records).Error
-	if err != nil {
-		return nil, fmt.Errorf("failed to query deployment history: %w", err)
-	}
-	return records, nil
-}
-
 // GetAppDeploymentHistory returns the deployment history for an app by app ID.
-
-// ---------- 11. DetectEnv ----------
-
-func (b *Bridge) DetectEnv(ctx context.Context, level int, ports []int, services []string) (interface{}, error) {
-	env := map[string]interface{}{}
-
-	// Level 1: OS info
-	if level >= 1 {
-		if out, err := b.Executor.RunCommand(ctx, "uname -a"); err == nil {
-			env["os"] = strings.TrimSpace(out)
-		}
-		if out, err := b.Executor.RunCommand(ctx, "cat /etc/os-release 2>/dev/null | head -5"); err == nil {
-			env["os_release"] = strings.TrimSpace(out)
-		}
-	}
-
-	// Level 2: Docker info
-	if level >= 2 {
-		if out, err := b.Executor.RunCommand(ctx, "docker version --format '{{.Server.Version}}' 2>/dev/null"); err == nil {
-			env["docker_version"] = strings.TrimSpace(out)
-		}
-		if out, err := b.Executor.RunCommand(ctx, "docker info --format '{{.NCPU}}' 2>/dev/null"); err == nil {
-			env["docker_cpus"] = strings.TrimSpace(out)
-		}
-		if out, err := b.Executor.RunCommand(ctx, "docker info --format '{{.MemTotal}}' 2>/dev/null"); err == nil {
-			env["docker_memory"] = strings.TrimSpace(out)
-		}
-	}
-
-	// Level 3: Port checks
-	if level >= 3 {
-		portResults := map[string]bool{}
-		for _, p := range ports {
-			cmd := fmt.Sprintf("ss -tlnp 2>/dev/null | grep ':%d ' || true", p)
-			if out, err := b.Executor.RunCommand(ctx, cmd); err == nil && strings.TrimSpace(out) != "" {
-				portResults[fmt.Sprintf("%d", p)] = true
-			} else {
-				portResults[fmt.Sprintf("%d", p)] = false
-			}
-		}
-		env["ports"] = portResults
-	}
-
-	// Level 4: Service checks
-	if level >= 4 {
-		svcResults := map[string]bool{}
-		for _, svc := range services {
-			svc = strings.TrimSpace(svc)
-			if svc == "" {
-				continue
-			}
-			cmd := fmt.Sprintf("timeout 2 bash -c 'echo > /dev/tcp/%s' 2>/dev/null && echo ok || echo fail", strings.TrimPrefix(svc, "tcp://"))
-			if out, err := b.Executor.RunCommand(ctx, cmd); err == nil && strings.Contains(out, "ok") {
-				svcResults[svc] = true
-			} else {
-				svcResults[svc] = false
-			}
-		}
-		env["services"] = svcResults
-	}
-
-	return env, nil
-}
-
-// ---------- 12. HealthCheck ----------
-
-func (b *Bridge) HealthCheck(ctx context.Context, target, healthType string) (interface{}, error) {
-	cfg := deployer.HealthCheckConfig{
-		Type:     defaultVal(healthType, "http"),
-		Target:   target,
-		Timeout:  5 * time.Second,
-		Interval: 3 * time.Second,
-		Retries:  3,
-	}
-
-	start := time.Now()
-	err := b.d().HealthCheck(ctx, cfg)
-	elapsed := time.Since(start)
-
-	if err != nil {
-		return map[string]interface{}{
-			"status":   "unhealthy",
-			"target":   target,
-			"type":     cfg.Type,
-			"error":    err.Error(),
-			"duration": elapsed.String(),
-		}, nil
-	}
-
-	return map[string]interface{}{
-		"status":   "healthy",
-		"target":   target,
-		"type":     cfg.Type,
-		"duration": elapsed.String(),
-	}, nil
-}
 
 
 // ---------- 16. CreateCredential ----------
 
 
-// ---------- 33. CheckDeployReadiness ----------
-
-func (b *Bridge) CheckDeployReadiness(ctx context.Context, appConfig map[string]interface{}) (interface{}, error) {
-	checks := map[string]interface{}{}
-
-	// Check Docker availability
-	if out, err := b.Executor.RunCommand(ctx, "docker version --format '{{.Server.Version}}' 2>/dev/null"); err == nil {
-		checks["docker"] = map[string]interface{}{"available": true, "version": strings.TrimSpace(out)}
-	} else {
-		checks["docker"] = map[string]interface{}{"available": false, "error": err.Error()}
-	}
-
-	// Check if the requested port is free
-	if portsRaw, ok := appConfig["ports"]; ok {
-		portsStr := fmt.Sprintf("%v", portsRaw)
-		hostPort := portsStr
-		if idx := strings.Index(portsStr, ":"); idx >= 0 {
-			hostPort = portsStr[:idx]
-		}
-		cmd := fmt.Sprintf("ss -tlnp 2>/dev/null | grep ':%s ' || true", hostPort)
-		if out, err := b.Executor.RunCommand(ctx, cmd); err == nil && strings.TrimSpace(out) == "" {
-			checks["port"] = map[string]interface{}{"port": hostPort, "available": true}
-		} else {
-			checks["port"] = map[string]interface{}{"port": hostPort, "available": false, "in_use_by": strings.TrimSpace(out)}
-		}
-	}
-
-	overallReady := true
-	for _, v := range checks {
-		if m, ok := v.(map[string]interface{}); ok {
-			if avail, ok := m["available"].(bool); ok && !avail {
-				overallReady = false
-			}
-		}
-	}
-
-	return map[string]interface{}{
-		"ready":  overallReady,
-		"checks": checks,
-	}, nil
-}
-
-// ---------- 34. BatchDeploy ----------
-
-// BatchDeploy deploys multiple applications using the configured strategy.
-// It accepts the legacy []map[string]interface{} parameter (backward compatible)
-// and defaults to sequential strategy.
-func (b *Bridge) BatchDeploy(ctx context.Context, apps []map[string]interface{}) (interface{}, error) {
-	config := mcp.BatchDeployConfig{
-		Apps:     apps,
-		Strategy: mcp.StrategySequential, // default
-	}
-	d := &bridgeDeployer{bridge: b}
-	result := executeBatchDeploy(ctx, d, config)
-	return result, nil
-}
-
-// BatchDeployWithConfig deploys multiple applications using the given configuration.
-func (b *Bridge) BatchDeployWithConfig(ctx context.Context, config mcp.BatchDeployConfig) (*mcp.BatchDeployResult, error) {
-	d := &bridgeDeployer{bridge: b}
-	result := executeBatchDeploy(ctx, d, config)
-	return result, nil
-}
-
-
 // ---------- 38. BatchDNS ----------
 
-
-// ---------- 39. CheckSystemUpdate ----------
-// Implemented in upgrade.go
-
-// PerformSystemUpdate delegates to UpgradeService.
-func (b *Bridge) PerformSystemUpdate(ctx context.Context) (interface{}, error) {
-	if b.UpgradeSvc == nil {
-		return nil, fmt.Errorf("upgrade service not initialized")
-	}
-	return b.UpgradeSvc.PerformUpgrade(ctx, "latest")
-}
-
-// ---------- 40. HealContainer ----------
-
-func (b *Bridge) HealContainer(ctx context.Context, containerName string) (interface{}, error) {
-	h := b.getHealer()
-	result, err := h.CheckAndHeal(ctx, containerName)
-	if err != nil {
-		return nil, fmt.Errorf("heal failed for %s: %w", containerName, err)
-	}
-
-	// If healer determined a rollback is needed, execute it automatically
-	if result.Action == "rolled_back" && h.GetConfig().AutoRollback {
-		slog.Info("healer: initiating auto-rollback", "container", containerName, "reason", result.Reason)
-		_, rollbackErr := b.RollbackWithType(ctx, containerName, "", "auto_rollback")
-		if rollbackErr != nil {
-			slog.Error("healer: auto-rollback failed", "container", containerName, "error", rollbackErr)
-			result.Action = "rollback_failed"
-			result.Reason = fmt.Sprintf("%s (rollback failed: %v)", result.Reason, rollbackErr)
-		} else {
-			result.NewState = "rolled_back"
-			slog.Info("healer: auto-rollback succeeded", "container", containerName)
-		}
-	}
-
-	return result, nil
-}
-
-
-// ---------- helpers ----------
-
-// saveDeploymentRecord persists a deployment attempt to the database.
-func (b *Bridge) saveDeploymentRecord(ctx context.Context, cfg mcp.DeployConfig, status string, pfResult *PreflightResult) {
-	if b.DB == nil {
-		return
-	}
-
-	// Serialize config snapshot for rollback support
-	snapshotJSON, _ := json.Marshal(cfg)
-
-	// Resolve app_id from app_name if possible
-	var appID string
-	if cfg.AppName != "" {
-		var app model.App
-		if err := b.DB.Select("id").Where("name = ?", cfg.AppName).First(&app).Error; err == nil {
-			appID = app.ID
-		}
-	}
-
-	record := &model.DeploymentRecord{
-		ID:             generateID(),
-		TenantID:       "tenant-default",
-		ServerID:       cfg.ServerID,
-		AppName:        cfg.AppName,
-		AppID:          appID,
-		ContainerName:  cfg.ContainerName,
-		Image:          cfg.Image,
-		DeployType:     "deploy",
-		ConfigSnapshot: string(snapshotJSON),
-		Status:         status,
-		CreatedAt:      time.Now(),
-		UpdatedAt:      time.Now(),
-	}
-	if pfResult != nil {
-		record.PreflightCode = string(pfResult.Code)
-		record.PreflightMessage = pfResult.Message
-		checksJSON, _ := json.Marshal(pfResult.Checks)
-		record.PreflightChecks = string(checksJSON)
-	}
-	if err := b.DB.Create(record).Error; err != nil {
-		slog.Error("failed to save deployment record", "error", err)
-	}
-}
 
 // generateID returns a unique deployment record ID.
 func generateID() string {
@@ -1159,17 +396,6 @@ func logPreflightResult(containerName string, result *PreflightResult) {
 	}
 }
 
-// GetLatestDeploymentRecord returns the most recent deployment record for a container.
-func (b *Bridge) GetLatestDeploymentRecord(ctx context.Context, containerName string) (*model.DeploymentRecord, error) {
-	var record model.DeploymentRecord
-	err := b.DB.Where("container_name = ?", containerName).
-		Order("created_at DESC").
-		First(&record).Error
-	if err != nil {
-		return nil, err
-	}
-	return &record, nil
-}
 
 func toString(v interface{}) string {
 	if v == nil {
@@ -1216,519 +442,7 @@ func defaultVal(val, def string) string {
 	return val
 }
 
-// ---------- 45. TriggerCIBuild ----------
-
-func (b *Bridge) TriggerCIBuild(ctx context.Context, providerType, repo, branch string) (interface{}, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-
-	var provider model.Provider
-	err := b.DB.Where("type = ? AND enabled = ?", "cicd-"+providerType, true).First(&provider).Error
-	if err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"message": fmt.Sprintf("no enabled CI/CD provider found for type: %s", providerType),
-		}, nil
-	}
-
-	var cfg struct {
-		Token   string `json:"token"`
-		Owner   string `json:"owner"`
-		BaseURL string `json:"base_url"`
-	}
-	if err := json.Unmarshal([]byte(provider.Config), &cfg); err != nil {
-		return nil, fmt.Errorf("failed to parse CI/CD provider config: %w", err)
-	}
-
-	switch providerType {
-	case "github-actions":
-		gh := cicd.NewGitHubActionsProvider(cfg.Token, cfg.Owner)
-		runID, err := gh.TriggerBuild(ctx, repo, branch)
-		if err != nil {
-			return map[string]interface{}{
-				"status":  "error",
-				"message": err.Error(),
-			}, nil
-		}
-		return map[string]interface{}{
-			"status":   "triggered",
-			"run_id":   runID,
-			"repo":     repo,
-			"branch":   branch,
-			"provider": providerType,
-		}, nil
-	case "gitea":
-		gt := cicd.NewGiteaActionsProvider(cfg.Token, cfg.Owner, cfg.BaseURL)
-		runID, err := gt.TriggerBuild(ctx, repo, branch)
-		if err != nil {
-			return map[string]interface{}{
-				"status":  "error",
-				"message": err.Error(),
-			}, nil
-		}
-		return map[string]interface{}{
-			"status":   "triggered",
-			"run_id":   runID,
-			"repo":     repo,
-			"branch":   branch,
-			"provider": providerType,
-		}, nil
-	default:
-		return nil, fmt.Errorf("unsupported CI/CD provider type: %s", providerType)
-	}
-}
-
-// ========== Kubernetes Cluster Management ==========
-
-// CreateCluster creates a new Kubernetes cluster.
-
-
-// ========== Kubernetes Deployment Operations ==========
-
-// K8sDeploy deploys an application to a Kubernetes cluster.
-func (b *Bridge) K8sDeploy(ctx context.Context, clusterID string, app *mcp.K8sDeployConfig) error {
-	if b.DB == nil {
-		return fmt.Errorf("database not available")
-	}
-
-	cluster, err := model.GetClusterWithSecrets(clusterID)
-	if err != nil {
-		return fmt.Errorf("cluster not found: %w", err)
-	}
-
-	k8sProvider, err := server.NewK8sProvider(cluster)
-	if err != nil {
-		return fmt.Errorf("failed to create k8s provider: %w", err)
-	}
-
-	deployCfg := &server.K8sDeployConfig{
-		Name:      app.Name,
-		Image:     app.Image,
-		Replicas:  app.Replicas,
-		Ports:     app.Ports,
-		EnvVars:   app.EnvVars,
-		Labels:    app.Labels,
-		Namespace: app.Namespace,
-	}
-
-	return k8sProvider.Deploy(ctx, deployCfg)
-}
-
-// K8sListDeployments lists deployments in a Kubernetes cluster.
-func (b *Bridge) K8sListDeployments(ctx context.Context, clusterID string) (interface{}, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-
-	cluster, err := model.GetClusterWithSecrets(clusterID)
-	if err != nil {
-		return nil, fmt.Errorf("cluster not found: %w", err)
-	}
-
-	k8sProvider, err := server.NewK8sProvider(cluster)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create k8s provider: %w", err)
-	}
-
-	deployments, err := k8sProvider.ListDeployments(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list deployments: %w", err)
-	}
-
-	// Convert to serializable format
-	items := make([]map[string]interface{}, 0, len(deployments))
-	for _, d := range deployments {
-		replicas := int32(0)
-		if d.Spec.Replicas != nil {
-			replicas = *d.Spec.Replicas
-		}
-		items = append(items, map[string]interface{}{
-			"name":            d.Name,
-			"namespace":       d.Namespace,
-			"replicas":        replicas,
-			"available":       d.Status.AvailableReplicas,
-			"image":           d.Spec.Template.Spec.Containers[0].Image,
-			"created_at":      d.CreationTimestamp.Format("2006-01-02T15:04:05Z"),
-		})
-	}
-
-	return map[string]interface{}{
-		"status":  "success",
-		"total":   len(items),
-		"cluster": clusterID,
-		"items":   items,
-	}, nil
-}
-
-// K8sGetPods retrieves pods from a Kubernetes cluster.
-func (b *Bridge) K8sGetPods(ctx context.Context, clusterID, labelSelector string) (interface{}, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-
-	cluster, err := model.GetClusterWithSecrets(clusterID)
-	if err != nil {
-		return nil, fmt.Errorf("cluster not found: %w", err)
-	}
-
-	k8sProvider, err := server.NewK8sProvider(cluster)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create k8s provider: %w", err)
-	}
-
-	pods, err := k8sProvider.GetPods(ctx, labelSelector)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get pods: %w", err)
-	}
-
-	// Convert to serializable format
-	items := make([]map[string]interface{}, 0, len(pods))
-	for _, p := range pods {
-		var containers []map[string]string
-		for _, c := range p.Spec.Containers {
-			containers = append(containers, map[string]string{
-				"name":  c.Name,
-				"image": c.Image,
-			})
-		}
-
-		items = append(items, map[string]interface{}{
-			"name":              p.Name,
-			"namespace":         p.Namespace,
-			"status":            string(p.Status.Phase),
-			"pod_ip":            p.Status.PodIP,
-			"node":              p.Spec.NodeName,
-			"restart_count":     p.Status.ContainerStatuses[0].RestartCount,
-			"created_at":        p.CreationTimestamp.Format("2006-01-02T15:04:05Z"),
-			"containers":        containers,
-		})
-	}
-
-	return map[string]interface{}{
-		"status":  "success",
-		"total":   len(items),
-		"cluster": clusterID,
-		"items":   items,
-	}, nil
-}
-
-// ---------- PluginOps ----------
-
-// PluginOps handles plugin lifecycle operations (enable, disable, reload).
-func (b *Bridge) PluginOps(pluginID string, action string) (interface{}, error) {
-	if b.PluginMgr == nil {
-		return nil, fmt.Errorf("plugin manager not available")
-	}
-
-	ctx := context.Background()
-
-	switch action {
-	case "enable":
-		if err := b.PluginMgr.EnablePlugin(ctx, pluginID); err != nil {
-			return nil, err
-		}
-		return map[string]interface{}{
-			"status":    "success",
-			"plugin_id": pluginID,
-			"action":    "enabled",
-		}, nil
-
-	case "disable":
-		if err := b.PluginMgr.DisablePlugin(ctx, pluginID); err != nil {
-			return nil, err
-		}
-		return map[string]interface{}{
-			"status":    "success",
-			"plugin_id": pluginID,
-			"action":    "disabled",
-		}, nil
-
-	case "reload":
-		if err := b.PluginMgr.ReloadPlugin(ctx, pluginID); err != nil {
-			return nil, err
-		}
-		return map[string]interface{}{
-			"status":    "success",
-			"plugin_id": pluginID,
-			"action":    "reloaded",
-		}, nil
-
-	default:
-		return nil, fmt.Errorf("unknown plugin action: %s (valid: enable, disable, reload)", action)
-	}
-}
-
-// ---------- ListPlugins ----------
-
-// ListPlugins returns all plugins from DB, optionally filtered by provider.
-func (b *Bridge) ListPlugins(provider string) (interface{}, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-
-	plugins, err := model.ListPlugins("tenant-default", provider)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list plugins: %w", err)
-	}
-
-	// Enrich with registry descriptor info
-	result := make([]map[string]interface{}, 0, len(plugins))
-	for _, p := range plugins {
-		entry := map[string]interface{}{
-			"id":           p.ID,
-			"name":         p.Name,
-			"display_name": p.DisplayName,
-			"version":      p.Version,
-			"description":  p.Description,
-			"author":       p.Author,
-			"provider":     p.Provider,
-			"type":         p.Type,
-			"enabled":      p.Enabled,
-			"priority":     p.Priority,
-			"status":       p.Status,
-			"created_at":   p.CreatedAt,
-			"updated_at":   p.UpdatedAt,
-		}
-		if p.ErrorMsg != "" {
-			entry["error_msg"] = p.ErrorMsg
-		}
-		result = append(result, entry)
-	}
-
-	return map[string]interface{}{
-		"status": "success",
-		"total":  len(result),
-		"plugins": result,
-	}, nil
-}
-
-// ---------- GetPluginInfo ----------
-
-// GetPluginInfo returns detailed information about a specific plugin.
-func (b *Bridge) GetPluginInfo(pluginID string) (interface{}, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-
-	p, err := model.GetPlugin(pluginID)
-	if err != nil {
-		return nil, fmt.Errorf("plugin not found: %w", err)
-	}
-
-	result := map[string]interface{}{
-		"id":           p.ID,
-		"name":         p.Name,
-		"display_name": p.DisplayName,
-		"version":      p.Version,
-		"description":  p.Description,
-		"author":       p.Author,
-		"provider":     p.Provider,
-		"type":         p.Type,
-		"enabled":      p.Enabled,
-		"priority":     p.Priority,
-		"status":       p.Status,
-		"created_at":   p.CreatedAt,
-		"updated_at":   p.UpdatedAt,
-	}
-
-	if p.ErrorMsg != "" {
-		result["error_msg"] = p.ErrorMsg
-	}
-
-	// Check if instance is loaded
-	if b.PluginMgr != nil {
-		instance, err := b.PluginMgr.GetPluginInstance(pluginID)
-		if err == nil {
-			result["instance_loaded"] = true
-			result["instance_type"] = fmt.Sprintf("%T", instance)
-		} else {
-			result["instance_loaded"] = false
-		}
-	}
-
-	return map[string]interface{}{
-		"status": "success",
-		"plugin": result,
-	}, nil
-}
-
-// ---------- RegistryOps ----------
-
-// RegistryOps handles registry operations (login, push, list_tags, ping).
-func (b *Bridge) RegistryOps(registryID string, operation string, args map[string]interface{}) (interface{}, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-
-	// Load registry from DB
-	var reg model.Registry
-	if err := b.DB.Where("id = ?", registryID).First(&reg).Error; err != nil {
-		return nil, fmt.Errorf("registry not found: %w", err)
-	}
-
-	// Decrypt password
-	var row struct {
-		Password string `gorm:"column:password"`
-	}
-	if err := b.DB.Table("registries").Where("id = ?", registryID).Select("password").Take(&row).Error; err != nil {
-		return nil, fmt.Errorf("failed to load registry credentials: %w", err)
-	}
-	plainPassword := ""
-	if b.EncryptionKey != nil && row.Password != "" {
-		if decrypted, err := crypto.Decrypt(b.EncryptionKey, row.Password); err == nil {
-			plainPassword = decrypted
-		}
-	}
-
-	// Allow args to override registry fields (for inline auth)
-	regURL := reg.URL
-	regUser := reg.Username
-	regPass := plainPassword
-	if args != nil {
-		if v, ok := args["registry_url"].(string); ok && v != "" {
-			regURL = v
-		}
-		if v, ok := args["username"].(string); ok && v != "" {
-			regUser = v
-		}
-		if v, ok := args["password"].(string); ok && v != "" {
-			regPass = v
-		}
-	}
-
-	// Create registry provider
-	provider, err := registry.NewRegistryProvider(reg.Provider, regURL, regUser, regPass)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create registry provider: %w", err)
-	}
-
-	ctx := context.Background()
-
-	switch operation {
-	case "login":
-		if err := provider.Login(ctx); err != nil {
-			return nil, fmt.Errorf("registry login failed: %w", err)
-		}
-		return map[string]interface{}{
-			"status":  "success",
-			"message": fmt.Sprintf("successfully authenticated with registry %s", reg.Name),
-			"registry_id": reg.ID,
-		}, nil
-
-	case "push":
-		localImage, _ := args["local_image"].(string)
-		remoteTag, _ := args["remote_tag"].(string)
-		if localImage == "" {
-			return nil, fmt.Errorf("local_image is required")
-		}
-		if err := provider.Push(ctx, localImage, remoteTag); err != nil {
-			return nil, fmt.Errorf("push failed: %w", err)
-		}
-		result := map[string]interface{}{
-			"status":      "success",
-			"message":     "image pushed successfully",
-			"local_image": localImage,
-			"registry_id": reg.ID,
-		}
-		if remoteTag != "" {
-			result["remote_tag"] = remoteTag
-		}
-		return result, nil
-
-	case "list_tags":
-		repo, _ := args["repository"].(string)
-		if repo == "" {
-			return nil, fmt.Errorf("repository is required")
-		}
-		tags, err := provider.ListTags(ctx, repo)
-		if err != nil {
-			return nil, fmt.Errorf("list tags failed: %w", err)
-		}
-		return map[string]interface{}{
-			"status":     "success",
-			"repository": repo,
-			"tags":       tags,
-			"total":      len(tags),
-		}, nil
-
-	case "ping":
-		if err := provider.Ping(ctx); err != nil {
-			return map[string]interface{}{
-				"status":  "unreachable",
-				"message": err.Error(),
-				"registry_id": reg.ID,
-			}, nil
-		}
-		return map[string]interface{}{
-			"status":      "reachable",
-			"message":     "registry is accessible",
-			"registry_id": reg.ID,
-		}, nil
-
-	default:
-		return nil, fmt.Errorf("unknown registry operation: %s", operation)
-	}
-}
 
 // ---------- GetServersByTags ----------
 
 
-// ---------- 46. GetCIBuildStatus ----------
-
-func (b *Bridge) GetCIBuildStatus(ctx context.Context, providerType, runID string) (interface{}, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-
-	var provider model.Provider
-	err := b.DB.Where("type = ? AND enabled = ?", "cicd-"+providerType, true).First(&provider).Error
-	if err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"message": fmt.Sprintf("no enabled CI/CD provider found for type: %s", providerType),
-		}, nil
-	}
-
-	var cfg struct {
-		Token   string `json:"token"`
-		Owner   string `json:"owner"`
-		BaseURL string `json:"base_url"`
-	}
-	if err := json.Unmarshal([]byte(provider.Config), &cfg); err != nil {
-		return nil, fmt.Errorf("failed to parse CI/CD provider config: %w", err)
-	}
-
-	switch providerType {
-	case "github-actions":
-		gh := cicd.NewGitHubActionsProvider(cfg.Token, cfg.Owner)
-		status, err := gh.GetBuildStatus(ctx, runID)
-		if err != nil {
-			return map[string]interface{}{
-				"status":  "error",
-				"message": err.Error(),
-			}, nil
-		}
-		return map[string]interface{}{
-			"status":   "success",
-			"build":    status,
-			"provider": providerType,
-		}, nil
-	case "gitea":
-		gt := cicd.NewGiteaActionsProvider(cfg.Token, cfg.Owner, cfg.BaseURL)
-		status, err := gt.GetBuildStatus(ctx, runID)
-		if err != nil {
-			return map[string]interface{}{
-				"status":  "error",
-				"message": err.Error(),
-			}, nil
-		}
-		return map[string]interface{}{
-			"status":   "success",
-			"build":    status,
-			"provider": providerType,
-		}, nil
-	default:
-		return nil, fmt.Errorf("unsupported CI/CD provider type: %s", providerType)
-	}
-}
